feat(components/TextEditor): update TextEditor
