+++ components/molecules/TextEditor/styled.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const TextEditorWrapper = styled.div`
+  margin: 0 auto;
+
+  .ql-editor {
+    min-height: 320px;
+  }
+`;
