+++ package.json
@@ -46,6 +46,7 @@
     "cmdk": "^0.2.0",
     "date-fns": "^3.0.6",
     "embla-carousel-react": "^8.0.0-rc18",
+    "lodash": "4.17.21",
     "lucide-react": "^0.304.0",
     "next": "14.0.4",
     "next-themes": "^0.2.1",
@@ -62,14 +63,15 @@
     "react-quill": "^2.0.0",
     "react-resizable-panels": "^1.0.7",
     "sonner": "^1.3.1",
+    "stacktrace-js": "2.0.2",
+    "styled-components": "^6.1.8",
     "tailwind-merge": "^2.2.0",
     "tailwindcss-animate": "^1.0.7",
     "vaul": "^0.8.0",
-    "zod": "^3.22.4",
-    "stacktrace-js": "2.0.2",
-    "lodash": "4.17.21"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@types/lodash": "4.14.178",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
@@ -79,7 +81,6 @@
     "eslint-plugin-unused-imports": "^3.0.0",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
-    "typescript": "^5",
-    "@types/lodash": "4.14.178"
+    "typescript": "^5"
   }
 }
